+++ Source/Framework/Constants/DataStorage/UiMapType.cs
@@ -3,7 +3,7 @@
 
 namespace Framework.Constants;
 
-public enum UiMapType : uint
+public enum UiMapType : byte
 {
     Cosmic = 0,
     World = 1,
